+++ db/flush_job.h
@@ -62,7 +62,7 @@ class FlushJob {
            JobContext* job_context, LogBuffer* log_buffer,
            Directory* db_directory, Directory* output_file_directory,
            CompressionType output_compression, Statistics* stats,
-           EventLogger* event_logger);
+           EventLogger* event_logger, bool measure_io_stats);
 
   ~FlushJob();
 
@@ -93,6 +93,7 @@ class FlushJob {
   Statistics* stats_;
   EventLogger* event_logger_;
   TableProperties table_properties_;
+  bool measure_io_stats_;
 };
 
 }  // namespace rocksdb